Fix refreshing access token after expiration
The refresh endpoint that duplicacy is using (https://duplicacy.com/dropbox_refresh) doesn't appear to refresh the token if there is an existing access token passed in the body of the request (regardless of its `Expiry` value). This forces it to be blank before performing the refresh.

Alternatively, change https://duplicacy.com/dropbox_refresh to refresh if the token `Expiry` is in the past.