Close HTTP response when retrying on errors.
This attempts to fix the problem described in https://forum.duplicacy.com/t/cannot-check-chunks-socket-too-many-open-files/7385 which seems to occur when the Dropbox client retries on certain errors (401, 429, 500). In these cases, the client retries with another HTTP request, but it fails to close the existing one. Eventually, this leaks sockets and the process runs out of sockets.